# internal/webapi/models/login.go
package models

// @Description Confirmation token
type NewLogin struct {
	Username string `json:"email" example:"johndoe@mail.com" mapper:"username"`
	Password string `json:"password" mapper:"password"`
} //@name NewLogin

// @Description Login
type Login struct {
	AccessToken  string `json:"access_token" example:"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..." mapper:"access_token"`
	TokenType    string `json:"token_type" example:"Bearer" mapper:"token_type"`
	ExpiresIn    int    `json:"expires_in" example:"1800" mapper:"expires_in"`
	RefreshToken string `json:"refresh_token" example:"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..." mapper:"refresh_token"`
} //@name Login